🧪 test: update home key_event_handler test
